V1 hardware emulation fix for direct3d
The V1 emulation in hardware mode fails glide sdk test# 11, indicating
an issue with the framebuffer, possible in the writes. However, a
portion of the framebuffer did not fail the test. This issue produced
severe artifacts in direct3d6 games and higher, but was apparent in
direct3d5 games and even in DOS glide. The attached patch is a
workaround to this issue. This is not fully tested yet, but Jedi Knight
(d3d mode) has no artifacts with this patch and also the on-screen text
is clear to read, unlike without the patch.

This patch is for testing and is specific to V1 hardware emulation at
640x480, although it shouldn't affect dosbox outside of the V1
emulation. I posted here as soon as I had a successful result.

http://www.vogons.org/viewtopic.php?f=32&t=40113#p369437
v1_d3d_workaround-2.diff